+++ exercises/practice/yacht/yacht.h
@@ -1,8 +1,6 @@
 #ifndef YACHT_H
 #define YACHT_H
 
-#include <stdint.h>
-
 typedef enum {
    ONES,
    TWOS,
@@ -19,7 +17,7 @@ typedef enum {
 } category_t;
 
 typedef struct {
-   uint8_t faces[5];
+   int faces[5];
 } dice_t;
 
 int score(dice_t dice, category_t category);
